+++ src/snowflake/snowpark/_internal/analyzer/analyzer_utils.py
@@ -185,6 +185,7 @@ WITH = "WITH "
 DEFAULT_ON_NULL = " DEFAULT ON NULL "
 ANY = " ANY "
 ICEBERG = " ICEBERG "
+RENAME_FIELDS = " RENAME FIELDS"
 
 TEMPORARY_STRING_SET = frozenset(["temporary", "temp"])
 
@@ -1117,13 +1118,16 @@ def rank_related_function_expression(
     )
 
 
-def cast_expression(child: str, datatype: DataType, try_: bool = False) -> str:
+def cast_expression(
+    child: str, datatype: DataType, try_: bool = False, is_rename: bool = False
+) -> str:
     return (
         (TRY_CAST if try_ else CAST)
         + LEFT_PARENTHESIS
         + child
         + AS
         + convert_sp_to_sf_type(datatype)
+        + (RENAME_FIELDS if is_rename else "")
         + RIGHT_PARENTHESIS
     )